@@ -3,15 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@chakra-ui/react": "^2.5.1",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "axios": "^0.19.2",
     "bootstrap": "^4.4.1",
+    "framer-motion": "^10.2.4",
     "jsonwebtoken": "^8.5.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "^3.4.3"
   },
   "scripts": {
@@ -20,7 +24,6 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "start-both": "concurrently \"cd ../ && cd backend/ && PORT=3001 nodemon index.js\" \"PORT=3000 react-scripts start\""
-
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -30,5 +33,8 @@
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "devDependencies": {
+    "@babel/preset-env": "^7.20.2"
+  }
 }
