@@ -1,30 +1,33 @@
 import React, { useState } from "react";
 import { useHistory } from "react-router-dom";
 
-/** Login form.
+/** Signup form.
  *
  * Shows form and manages update to state on changes.
  * On submission:
- * - calls login function prop
+ * - calls signup function prop
  * - redirects to /companies route
  *
- * Routes -> LoginForm -> Alert
- * Routed as /login
+ * Routes -> SignupForm -> Alert
+ * Routed as /signup
  */
 
-function LoginForm({ login }) {
+function SignupForm({ signup }) {
   const history = useHistory();
   const [formData, setFormData] = useState({
     username: "",
     password: "",
+    firstName: "",
+    lastName: "",
+    email: "",
   });
   const [formErrors, setFormErrors] = useState([]);
 
   console.debug(
-      "LoginForm",
-      "login=", typeof login,
+      "SignupForm",
+      "signup=", typeof signup,
       "formData=", formData,
-      "formErrors", formErrors,
+      "formErrors=", formErrors,
   );
 
   /** Handle form submit:
@@ -34,7 +37,7 @@ function LoginForm({ login }) {
 
   async function handleSubmit(evt) {
     evt.preventDefault();
-    let result = await login(formData);
+    let result = await signup(formData);
     if (result.success) {
       history.push("/companies");
     } else {
@@ -45,14 +48,13 @@ function LoginForm({ login }) {
   /** Update form data field */
   function handleChange(evt) {
     const { name, value } = evt.target;
-    setFormData(l => ({ ...l, [name]: value }));
+    setFormData(data => ({ ...data, [name]: value }));
   }
 
   return (
-      <div className="LoginForm">
+      <div className="SignupForm">
         <div className="container col-md-6 offset-md-3 col-lg-4 offset-lg-4">
-          <h3 className="mb-3">Log In</h3>
-
+          <h2 className="mb-3">Sign Up</h2>
           <div className="card">
             <div className="card-body">
               <form onSubmit={handleSubmit}>
@@ -63,8 +65,6 @@ function LoginForm({ login }) {
                       className="form-control"
                       value={formData.username}
                       onChange={handleChange}
-                      autoComplete="username"
-                      required
                   />
                 </div>
                 <div className="form-group">
@@ -75,16 +75,45 @@ function LoginForm({ login }) {
                       className="form-control"
                       value={formData.password}
                       onChange={handleChange}
-                      autoComplete="current-password"
-                      required
+                  />
+                </div>
+
+                <div className="form-group">
+                  <label>First name</label>
+                  <input
+                      name="firstName"
+                      className="form-control"
+                      value={formData.firstName}
+                      onChange={handleChange}
+                  />
+                </div>
+                <div className="form-group">
+                  <label>Last name</label>
+                  <input
+                      name="lastName"
+                      className="form-control"
+                      value={formData.lastName}
+                      onChange={handleChange}
+                  />
+                </div>
+                <div className="form-group">
+                  <label>Email</label>
+                  <input
+                      type="email"
+                      name="email"
+                      className="form-control"
+                      value={formData.email}
+                      onChange={handleChange}
                   />
                 </div>
 
                 {formErrors.length
-                    ? console.log('error', formErrors)
-                    : null}
+                    ? console.log('error',formErrors)
+                    : null
+                }
 
                 <button
+                    type="submit"
                     className="btn btn-primary float-right"
                     onSubmit={handleSubmit}
                 >
@@ -98,4 +127,4 @@ function LoginForm({ login }) {
   );
 }
 
-export default LoginForm;
+export default SignupForm;
